feat: move graphene types to a types file
